+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **Added:** `itron::abi::{TA_TPRI, TA_WMUL, TA_CLR, TWF_*, TTW_*}`
 - **Added:** `itron::semaphore::Semaphore[Ref]`
 - **Added:** `itron::abi::{*_mbf}` (message buffers) and Cargo feature `messagebuf`
+- **Added:** `itron::abi::{*_ovr}` (overrun handlers) and Cargo feature `ovrhdr`
 
 ## 0.1.0 - 2021-06-15